+++ presentation/src/main/java/com/tdtd/presentation/ui/detail/DetailSharedBottomSheetFragment.kt
@@ -18,9 +18,7 @@ import com.google.firebase.analytics.ktx.analytics
 import com.google.firebase.ktx.Firebase
 import com.tdtd.presentation.R
 import com.tdtd.presentation.databinding.DetailAdminBottomSheetBinding
-import com.tdtd.presentation.util.getNavigationResult
-import com.tdtd.presentation.util.onThrottleClick
-import com.tdtd.presentation.util.showToast
+import com.tdtd.presentation.util.*
 
 class DetailSharedBottomSheetFragment : BottomSheetDialogFragment() {
 
@@ -58,19 +56,10 @@ class DetailSharedBottomSheetFragment : BottomSheetDialogFragment() {
     }
 
     private fun deleteRoom() {
-        getNavigationResult<String>(R.id.detailSharedBottomSheetFragment, "detail") {
-            findNavController().navigateUp()
+        getNavigationResult<String>(R.id.detailSharedBottomSheetFragment, "detail_delete_room") { text ->
             findNavController().popBackStack()
-            detailViewModel.deleteRoom(it)
-        }
-
-        detailViewModel.deleteRoomValue.observe(viewLifecycleOwner) { text ->
-            if (text.isNotEmpty()) {
-                requireActivity().showToast(
-                    getString(R.string.toast_delete_room_success),
-                    requireView()
-                )
-            }
+            requireActivity().showToast(text, requireView())
+            findNavController().navigateSafeUp(findNavController().currentDestination!!.id)
         }
     }
