+++ presentation/src/main/java/com/tdtd/presentation/ui/main/MainAdapter.kt
@@ -39,6 +39,8 @@ class MainAdapter(
 
         fun bind(data: Room) {
             binding.setVariable(BR.room, data)
+            binding.executePendingBindings()
+
             binding.container.setOnClickListener {
                 onClick(data)
             }
