+++ presentation/src/main/java/com/tdtd/presentation/util/AppExt.kt
@@ -3,7 +3,6 @@ package com.tdtd.presentation.util
 import android.app.Activity
 import android.content.Context
 import android.content.res.Resources
-import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import android.os.SystemClock
@@ -12,10 +11,6 @@ import android.view.*
 import android.view.inputmethod.InputMethodManager
 import android.widget.PopupWindow
 import android.widget.TextView
-import androidx.annotation.IdRes
-import androidx.navigation.NavController
-import androidx.navigation.NavOptions
-import androidx.navigation.Navigator
 import com.tdtd.presentation.R
 import kotlinx.android.synthetic.main.layout_toast.view.*
 import java.util.concurrent.TimeUnit
@@ -143,17 +138,4 @@ fun View.clickWithDebounce(debounceTime: Long = 300, action: () -> Unit) {
             lastClickTime = SystemClock.elapsedRealtime()
         }
     })
-}
-
-fun NavController.navigateSafe(
-    @IdRes resId: Int,
-    args: Bundle? = null,
-    navOptions: NavOptions? = null,
-    navExtras: Navigator.Extras? = null
-) {
-    val action = currentDestination?.getAction(resId) ?: graph.getAction(resId)
-
-    if (action != null && currentDestination?.id != action.destinationId) {
-        navigate(resId, args, navOptions, navExtras)
-    }
 }
